Cookbook Dependency imports can be marked as read-only
Ordinarily such an import would not yield any ActiveRecord objects if
the destination cookbook version already had its dependencies imported.
If an import is read-only it will yield the same ActiveRecord objects as
it would if the destination cookbook version had not had its
dependencies imported, but with the catch that the yielded objects are
read-only.